# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowStock API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Organization {
  id: number;
  name: string;
  createdAt: string;
}

export interface CreateOrganizationBody {
  name: string;
}

export interface Store {
  id: number;
  organizationId: number;
  name: string;
  address?: string;
  createdAt: string;
}

export interface CreateStoreBody {
  name: string;
  address?: string;
}

export interface InventoryItem {
  id: number;
  storeId?: number;
  name: string;
  stock: number;
  unit: string;
  minStock?: number;
  createdAt: string;
}

export interface CreateInventoryItemBody {
  storeId?: number;
  name: string;
  stock: number;
  unit: string;
  minStock?: number;
}

export interface UpdateInventoryItemBody {
  name?: string;
  stock?: number;
  unit?: string;
  minStock?: number;
}

export interface Sale {
  id: number;
  storeId?: number;
  menuItem: string;
  quantity: number;
  createdAt: string;
}

export interface CreateSaleBody {
  storeId?: number;
  menuItem: string;
  quantity: number;
}

export interface RecipeIngredient {
  ingredientName: string;
  amountPerServing: number;
}

export interface Recipe {
  id: number;
  menuItem: string;
  ingredients: RecipeIngredient[];
  organizationId?: number;
  storeId?: number;
  isCompanyRecipe: boolean;
  createdAt: string;
}

export interface CreateRecipeBody {
  menuItem: string;
  ingredients: RecipeIngredient[];
  storeId?: number;
}

export interface Recommendation {
  ingredientName: string;
  currentStock: number;
  unit: string;
  totalUsed: number;
  orderAmount: number;
}

export type DashboardSummaryTopMenuItemsItem = {
  menuItem: string;
  totalQty: number;
};

export interface DashboardSummary {
  totalInventoryItems: number;
  totalSales: number;
  totalRecipes: number;
  lowStockCount: number;
  reorderCount: number;
  topMenuItems: DashboardSummaryTopMenuItemsItem[];
}

export type ListInventoryParams = {
  storeId?: number;
};

export type ListSalesParams = {
  storeId?: number;
};

export type ListRecipesParams = {
  storeId?: number;
  organizationId?: number;
};

export type GetRecommendationsParams = {
  storeId?: number;
  organizationId?: number;
};

export type GetDashboardSummaryParams = {
  storeId?: number;
  organizationId?: number;
};