# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: FlowStock API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: organizations
    description: Organization management
  - name: stores
    description: Store management
  - name: inventory
    description: Inventory management
  - name: sales
    description: Sales tracking
  - name: recipes
    description: Recipe management
  - name: recommendations
    description: Reorder recommendations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List all organizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

  /organizations/{id}:
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /organizations/{id}/stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores for an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Add a store to an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

  /stores/{id}:
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Delete a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /organizations/{id}/recipes:
    get:
      operationId: listCompanyRecipes
      tags: [recipes]
      summary: List company-wide recipes for an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of company recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createCompanyRecipe
      tags: [recipes]
      summary: Create a company-wide recipe for an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"

  /organizations/{id}/recipes/{recipeId}:
    delete:
      operationId: deleteCompanyRecipe
      tags: [recipes]
      summary: Delete a company recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items, optionally scoped to a store
      parameters:
        - name: storeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  /inventory/{id}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales, optionally scoped to a store
      parameters:
        - name: storeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Log a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  /sales/{id}:
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List store-level recipes, optionally merged with company recipes
      parameters:
        - name: storeId
          in: query
          required: false
          schema:
            type: integer
        - name: organizationId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a store-level recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"

  /recipes/{id}:
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a store-level recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /recommendations:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: Get reorder recommendations, optionally scoped to a store
      parameters:
        - name: storeId
          in: query
          required: false
          schema:
            type: integer
        - name: organizationId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of reorder recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [recommendations]
      summary: Get dashboard summary stats, optionally scoped to a store
      parameters:
        - name: storeId
          in: query
          required: false
          schema:
            type: integer
        - name: organizationId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt

    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    Store:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - name
        - createdAt

    CreateStoreBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
      required:
        - name

    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        name:
          type: string
        stock:
          type: number
        unit:
          type: string
        minStock:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - stock
        - unit
        - createdAt

    CreateInventoryItemBody:
      type: object
      properties:
        storeId:
          type: integer
        name:
          type: string
        stock:
          type: number
        unit:
          type: string
        minStock:
          type: number
      required:
        - name
        - stock
        - unit

    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        stock:
          type: number
        unit:
          type: string
        minStock:
          type: number

    Sale:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        menuItem:
          type: string
        quantity:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - menuItem
        - quantity
        - createdAt

    CreateSaleBody:
      type: object
      properties:
        storeId:
          type: integer
        menuItem:
          type: string
        quantity:
          type: number
      required:
        - menuItem
        - quantity

    RecipeIngredient:
      type: object
      properties:
        ingredientName:
          type: string
        amountPerServing:
          type: number
      required:
        - ingredientName
        - amountPerServing

    Recipe:
      type: object
      properties:
        id:
          type: integer
        menuItem:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        organizationId:
          type: integer
        storeId:
          type: integer
        isCompanyRecipe:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - menuItem
        - ingredients
        - isCompanyRecipe
        - createdAt

    CreateRecipeBody:
      type: object
      properties:
        menuItem:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        storeId:
          type: integer
      required:
        - menuItem
        - ingredients

    Recommendation:
      type: object
      properties:
        ingredientName:
          type: string
        currentStock:
          type: number
        unit:
          type: string
        totalUsed:
          type: number
        orderAmount:
          type: number
      required:
        - ingredientName
        - currentStock
        - unit
        - totalUsed
        - orderAmount

    DashboardSummary:
      type: object
      properties:
        totalInventoryItems:
          type: integer
        totalSales:
          type: integer
        totalRecipes:
          type: integer
        lowStockCount:
          type: integer
        reorderCount:
          type: integer
        topMenuItems:
          type: array
          items:
            type: object
            properties:
              menuItem:
                type: string
              totalQty:
                type: number
            required:
              - menuItem
              - totalQty
      required:
        - totalInventoryItems
        - totalSales
        - totalRecipes
        - lowStockCount
        - reorderCount
        - topMenuItems