# lib/api-zod/src/generated/types/createInventoryItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowStock API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateInventoryItemBody {
  storeId?: number;
  name: string;
  stock: number;
  unit: string;
  minStock?: number;
}