# lib/api-zod/src/generated/types/createRecipeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowStock API specification
 * OpenAPI spec version: 0.2.0
 */
import type { RecipeIngredient } from "./recipeIngredient";

export interface CreateRecipeBody {
  menuItem: string;
  ingredients: RecipeIngredient[];
  storeId?: number;
}